Adjust example, move plugin to separate file
